+++ src/main/java/me/ajlane/geo/repast/colonisation/package-info.java
@@ -1,0 +1,16 @@
+/**
+ * Classes implementing models that represent land-cover colonisation.
+ *
+ * <p>
+ * Colonisation in this sense refers to the combined effect of seed dispersal, seed germination,
+ * seedling establishment, and seedling survival (Thompson, 2005).
+ * </p>
+ *
+ * <h3>References</h3>
+ * <p>
+ * Thompson, J. D. (2005). Plant Evolution in the Mediterranean. Oxford: Oxford University Press.
+ * </p>
+ *
+ * @author Andrew Lane
+ */
+package me.ajlane.geo.repast.colonisation;