+++ src/main/java/repast/model/agrosuccess/params/EnvrModelParams.java
@@ -1,8 +1,8 @@
 package repast.model.agrosuccess.params;
 
+import me.ajlane.geo.repast.colonisation.SeedDispersalParams;
+import me.ajlane.geo.repast.colonisation.SeedViabilityParams;
 import me.ajlane.geo.repast.fire.FireParams;
-import me.ajlane.geo.repast.seeddispersal.SeedDispersalParams;
-import me.ajlane.geo.repast.seeddispersal.SeedViabilityParams;
 import me.ajlane.geo.repast.soilmoisture.SoilMoistureParams;
 
 /**